2025_housekeeping_II: move manager/worker into domain exc into common
