+++ job_executor/adapter/datastore_api/__init__.py
@@ -13,8 +13,8 @@ from job_executor.adapter.datastore_api.models import (
     MaintenanceStatus,
     Operation,
 )
+from job_executor.common.exceptions import HttpRequestError, HttpResponseError
 from job_executor.config import environment
-from job_executor.exception import HttpRequestError, HttpResponseError
 
 DATASTORE_API_URL = environment.datastore_api_url
 DEFAULT_REQUESTS_TIMEOUT = (10, 60)  # (read timeout, connect timeout)
