+++ job_executor/adapter/pseudonym_service.py
@@ -1,8 +1,8 @@
 import requests
 from microdata_tools.validation.model.metadata import UnitIdType
 
+from job_executor.common.exceptions import HttpResponseError
 from job_executor.config import environment, secrets
-from job_executor.exception import HttpResponseError
 
 PSEUDONYM_SERVICE_URL = environment.pseudonym_service_url
 PSEUDONYM_SERVICE_API_KEY = secrets.pseudonym_service_api_key
