+++ job_executor/domain/worker/steps/dataset_pseudonymizer.py
@@ -9,7 +9,7 @@ from pyarrow import compute, dataset, parquet
 
 from job_executor.adapter import pseudonym_service
 from job_executor.adapter.local_storage.models.metadata import Metadata
-from job_executor.exception import BuilderStepError
+from job_executor.common.exceptions import BuilderStepError
 
 logger = logging.getLogger()
