+++ job_executor/domain/worker/steps/dataset_validator.py
@@ -3,8 +3,8 @@ from pathlib import Path
 
 from microdata_tools import validate_dataset, validate_metadata
 
+from job_executor.common.exceptions import BuilderStepError
 from job_executor.config import environment
-from job_executor.exception import BuilderStepError
 
 logger = logging.getLogger()
 WORKING_DIR = Path(environment.working_dir)
