+++ tests/unit/manager/test_manager.py
@@ -1,8 +1,8 @@
 import time
 from multiprocessing import Process
 
-from job_executor.manager import Manager
-from job_executor.worker.models import Worker
+from job_executor.domain.manager import Manager
+from job_executor.domain.worker.models import Worker
 
 
 def dummy():
