+++ tests/unit/worker/steps/test_dataset_pseudonymizer.py
@@ -9,8 +9,8 @@ from pyarrow import dataset, parquet
 
 from job_executor.adapter import pseudonym_service
 from job_executor.adapter.local_storage.models.metadata import Metadata
-from job_executor.exception import BuilderStepError
-from job_executor.worker.steps import dataset_pseudonymizer
+from job_executor.common.exceptions import BuilderStepError
+from job_executor.domain.worker.steps import dataset_pseudonymizer
 
 TABLE_SIZE = 1000
 UNIT_ID_INPUT = [f"i{count}" for count in range(TABLE_SIZE)]
