+++ tests/unit/worker/test_build_dataset_worker.py
@@ -11,7 +11,7 @@ from requests_mock import Mocker as RequestsMocker
 
 from job_executor.adapter.local_storage import INPUT_DIR
 from job_executor.config import environment
-from job_executor.worker.build_dataset_worker import run_worker
+from job_executor.domain.worker.build_dataset_worker import run_worker
 
 RSA_KEYS_DIRECTORY = Path(environment.rsa_keys_directory)
