+++ tests/unit/worker/test_build_metadata_worker.py
@@ -9,7 +9,7 @@ from requests_mock import Mocker as RequestsMocker
 
 from job_executor.adapter.local_storage import INPUT_DIR
 from job_executor.config import environment
-from job_executor.worker.build_metadata_worker import run_worker
+from job_executor.domain.worker.build_metadata_worker import run_worker
 from tests.unit.worker.test_build_dataset_worker import _create_rsa_public_key
 
 RSA_KEYS_DIRECTORY = Path(environment.rsa_keys_directory)
